Fixed all known issues with the Ram.
The Ram now doesn't turn to keep facing the player whilst charging, he keeps looking in the direction that he's charging.
Fixed an issue where if the Ram hit a sheep during it's charge, it could no longer target the player it was attempting to charge when the sheep was hit.